Move paired tests to the same folder
